# Significantly speedup javascript bindings tasks by skipping libbls tests and benchmarks