Example `Build JS`:
* Before: ~4m 35s
* After: ~1m 38s

Example `Build and Test on ubuntu-latest CPython 3.11`:
* Before: ~4m 52s
* After: ~1m 57s